fixup! Frontend: Extract UI Components
IWinLoseView parametrically derives from IButton
