+++ forge-gui/src/main/java/forge/gamemodes/limited/LimitedWinLoseController.java
@@ -1,8 +1,8 @@
 package forge.gamemodes.limited;
 
 import forge.frontend.components.widgets.IButton;
+import forge.frontend.components.widgets.IWinLoseView;
 import forge.game.GameView;
-import forge.gui.interfaces.IWinLoseView;
 import forge.model.FModel;
 import forge.player.GamePlayerUtil;
 import forge.util.Localizer;
