+++ forge-gui/src/main/java/forge/gamemodes/planarconquest/ConquestBattle.java
@@ -5,10 +5,10 @@ import java.util.Set;
 import forge.LobbyPlayer;
 import forge.deck.Deck;
 import forge.frontend.components.widgets.IButton;
+import forge.frontend.components.widgets.IWinLoseView;
 import forge.game.GameType;
 import forge.game.GameView;
 import forge.gui.interfaces.IGuiGame;
-import forge.gui.interfaces.IWinLoseView;
 import forge.item.PaperCard;
 import forge.util.Localizer;
