+++ forge-gui/src/main/java/forge/gamemodes/planarconquest/ConquestChaosBattle.java
@@ -7,6 +7,7 @@ import java.util.Set;
 import forge.LobbyPlayer;
 import forge.deck.Deck;
 import forge.frontend.components.widgets.IButton;
+import forge.frontend.components.widgets.IWinLoseView;
 import forge.game.GameType;
 import forge.game.GameView;
 import forge.gamemodes.planarconquest.ConquestPreferences.CQPref;
@@ -15,7 +16,6 @@ import forge.gamemodes.quest.QuestEventDuel;
 import forge.gamemodes.quest.QuestEventDuelManager;
 import forge.gamemodes.quest.QuestWorld;
 import forge.gui.interfaces.IGuiGame;
-import forge.gui.interfaces.IWinLoseView;
 import forge.item.PaperCard;
 import forge.localinstance.properties.ForgeConstants;
 import forge.model.FModel;
