+++ forge-gui/src/main/java/forge/gamemodes/planarconquest/ConquestController.java
@@ -29,6 +29,7 @@ import forge.deck.Deck;
 import forge.deck.DeckSection;
 import forge.deck.io.DeckStorage;
 import forge.frontend.components.widgets.IButton;
+import forge.frontend.components.widgets.IWinLoseView;
 import forge.game.GameRules;
 import forge.game.GameType;
 import forge.game.GameView;
@@ -39,7 +40,6 @@ import forge.gamemodes.quest.BoosterUtils;
 import forge.gui.FThreads;
 import forge.gui.GuiBase;
 import forge.gui.interfaces.IGuiGame;
-import forge.gui.interfaces.IWinLoseView;
 import forge.item.PaperCard;
 import forge.localinstance.properties.ForgeConstants;
 import forge.localinstance.properties.ForgePreferences.FPref;
